refactor: Store versions inside the same file
