+++ src/frformat/formats/code_commune_insee.py
@@ -3,8 +3,9 @@ from frformat.formats.code_commune_insee_frozenset import (
     CODES_COMMUNES_INSEE_COG_2023,
     CODES_COMMUNES_INSEE_COG_2024,
 )
-from frformat.set_format import INSEE_SOURCE, Millesime
+from frformat.set_format import INSEE_SOURCE
 from frformat.versioned_set import VersionedSet
+from frformat.versions import Millesime
 
 name = "Code commune INSEE"
 description = "Vérifie que le code commune correspond bien à un code commune INSEE pour un Code Officiel Géographique donné"
