+++ src/frformat/formats/numero_departement.py
@@ -1,6 +1,7 @@
 from frformat import set_format
-from frformat.set_format import INSEE_SOURCE, Millesime
+from frformat.set_format import INSEE_SOURCE
 from frformat.versioned_set import VersionedSet
+from frformat.versions import Millesime
 
 NUMEROS_DEPARTEMENTS_COG_2023 = frozenset(
     (
